Youtube_Chatbot_Backend/requirements.txt
fastapi
uvicorn
python-dotenv
torch
transformers
sentence-transformers
pinecone
pinecone-text
langchain
langchain-ollama
langchain-huggingface
httpx
pydantic
